LayoutTests/media/track/track-cue-empty-cue-text-expected.txt

getCueAsHTML() on an empty cue should return a valid document fragment.


** The getCueAsHTML() method should return a document fragment **
EXPECTED (fragment != 'null') OK

** The document fragment should have one child, an empty Text node **
EXPECTED (fragment.childNodes.length == '1') OK
EXPECTED (fragment.childNodes[0].length == '0') OK
EXPECTED (fragment.childNodes[0].constructor.name == 'Text') OK

END OF TEST
